requirements.txt
click==8.0.3
grako==3.99.9
lxml==4.8.0
pygeoif==0.7
pystac==1.3.0
python-dateutil==2.8.2
python-slugify==6.0.1
six==1.16.0
text-unidecode==1.3
